pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "chameleon-mcp"
version = "0.2.0"
description = "The shape-shifting MCP hub — morph() into any server at runtime. 3,000+ servers. No config. No restart."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Chameleon MCP Contributors" }]
keywords = [
    "mcp",
    "model-context-protocol",
    "claude",
    "anthropic",
    "ai-agents",
    "llm-tools",
    "fastmcp",
    "smithery",
    "mcp-server",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP",
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "certifi>=2024.0.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "respx>=0.21.0",
]

[project.scripts]
chameleon-mcp = "server:mcp.run"

[project.urls]
Homepage = "https://github.com/kaiser-data/chameleon-mcp"
Repository = "https://github.com/kaiser-data/chameleon-mcp"
Issues = "https://github.com/kaiser-data/chameleon-mcp/issues"
Documentation = "https://github.com/kaiser-data/chameleon-mcp/tree/main/docs"

[tool.hatch.build.targets.wheel]
packages = ["."]
include = ["server.py", "chameleon_mcp/*.py"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008", "SIM117"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
